apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai-enterprise-pipeline-api"
version = "0.1.0"
description = "FastAPI control plane for the CrewAI Enterprise Pipeline flagship build."
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiosqlite==0.21.0",
  "alembic==1.18.4",
  "asyncpg==0.31.0",
  "boto3==1.42.76",
  "crewai==1.12.2",
  "fastapi==0.135.2",
  "httpx==0.28.1",
  "Jinja2>=3.1",
  "opentelemetry-api~=1.34.0",
  "opentelemetry-sdk~=1.34.0",
  "opentelemetry-exporter-otlp~=1.34.0",
  "opentelemetry-instrumentation-fastapi~=0.55b0",
  "opentelemetry-instrumentation-httpx~=0.55b0",
  "opentelemetry-instrumentation-sqlalchemy~=0.55b0",
  "openpyxl>=3.1",
  "pdfplumber>=0.11",
  "prometheus-client>=0.23,<1",
  "PyJWT>=2.10",
  "pydantic==2.11.9",
  "pydantic-settings==2.10.1",
  "python-docx==1.2.0",
  "reportlab>=4.2",
  "arq>=0.26",
  "redis>=5.0,<6",
  "pgvector>=0.3",
  "sqlalchemy==2.0.48",
  "structlog>=25.4,<26",
  "uvicorn[standard]==0.42.0",
]

[project.optional-dependencies]
dev = [
  "pytest==9.0.2",
  "ruff==0.15.7",
]
embeddings = [
  "sentence-transformers>=3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["crewai_enterprise_pipeline_api"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
crewai_enterprise_pipeline_api = ["templates/*.j2"]
